Java source code
A simple java program that uses the random number generator to select a magic number. The user will then try to guess the number. The program will check to see if the guess is a match or if is too high/low.

Contributors Welcome! 
Needs work: The user only gets one guess. A loop would allow the user to keep guessing until they arrived at the magic number.